pyproject.toml
[tool.poetry]
name = "caddy"
version = "0.0.1"
description = ""
authors = ["Christian Heider Lindbjerg <chen@mapspeople.com>"]
readme = "README.md"
packages = [{ include = "caddy" }]


[tool.poetry.dependencies]
pytest = "*"
python = "^3.9"
warg = ">=1.4.1"
draugr = ">=1.2.0"
numpy = "^1.24.0"
jord = ">=0.8.2"
shapely = ">=2.0.3"
ezdxf = ">=1.2.0"
geopandas = ">=1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = ">=6.0.0"
black = ">=23.1.0"
pre-commit = ">=2.0.0"
furo = ">=2022.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
